refactor: table 멀티모듈로 분리
