+++ kitchenpos-common/src/main/resources/db/migration/clean_up_tables.sql
@@ -1,0 +1,18 @@
+SET REFERENTIAL_INTEGRITY FALSE;
+TRUNCATE TABLE MENU;
+ALTER TABLE MENU ALTER COLUMN id RESTART WITH 1;
+TRUNCATE TABLE MENU_GROUP;
+ALTER TABLE MENU_GROUP ALTER COLUMN id RESTART WITH 1;
+TRUNCATE TABLE MENU_PRODUCT;
+ALTER TABLE MENU_PRODUCT ALTER COLUMN seq RESTART WITH 1;
+TRUNCATE TABLE ORDERS;
+ALTER TABLE ORDERS ALTER COLUMN id RESTART WITH 1;
+TRUNCATE TABLE ORDER_LINE_ITEM;
+ALTER TABLE ORDER_LINE_ITEM ALTER COLUMN seq RESTART WITH 1;
+TRUNCATE TABLE ORDER_TABLE;
+ALTER TABLE ORDER_TABLE ALTER COLUMN id RESTART WITH 1;
+TRUNCATE TABLE PRODUCT;
+ALTER TABLE PRODUCT ALTER COLUMN id RESTART WITH 1;
+TRUNCATE TABLE TABLE_GROUP;
+ALTER TABLE TABLE_GROUP ALTER COLUMN id RESTART WITH 1;
+SET REFERENTIAL_INTEGRITY TRUE;
