+++ kitchenpos-common/src/test/java/kitchenpos/acceptance/TableGroupAcceptanceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.acceptance;
+package kitchenpos.acceptance;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -18,9 +18,9 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.stream.Stream;
 
-import static kitchenpos.order.acceptance.OrderAcceptanceTest.메뉴_등록_요청;
-import static kitchenpos.order.acceptance.OrderAcceptanceTest.주문_생성_요청;
-import static kitchenpos.table.acceptance.TableAcceptanceTest.테이블_생성_요청;
+import static kitchenpos.acceptance.OrderAcceptanceTest.메뉴_등록_요청;
+import static kitchenpos.acceptance.OrderAcceptanceTest.주문_생성_요청;
+import static kitchenpos.acceptance.TableAcceptanceTest.테이블_생성_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.DynamicTest.dynamicTest;
