+++ kitchenpos-table/src/main/java/kitchenpos/table/domain/NumberOfGuests.java
