+++ kitchenpos-table/src/main/java/kitchenpos/table/domain/TableEmptyChangedEvent.java
