+++ kitchenpos-table/src/main/java/kitchenpos/table/domain/TableGroupUngroupedEvent.java
