+++ kitchenpos-table/src/main/java/kitchenpos/table/dto/EmptyRequest.java
