+++ kitchenpos-table/src/main/java/kitchenpos/table/dto/NumberOfGuestsRequest.java
