+++ kitchenpos-table/src/main/java/kitchenpos/table/dto/TableGroupResponse.java
