+++ kitchenpos-table/src/test/java/kitchenpos/table/domain/OrderTableTest.java
@@ -1,9 +1,9 @@
 package kitchenpos.table.domain;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class OrderTableTest {
@@ -11,14 +11,14 @@ class OrderTableTest {
     @Test
     void createEmptyTable() {
         OrderTable orderTable = new OrderTable(0, true);
-        assertThat(orderTable.isEmpty()).isTrue();
+        Assertions.assertThat(orderTable.isEmpty()).isTrue();
     }
 
     @DisplayName("비어있지 않은 OrderTable을 생성한다.")
     @Test
     void createNotEmptyTable() {
         OrderTable orderTable = new OrderTable(5, false);
-        assertThat(orderTable.isEmpty()).isFalse();
+        Assertions.assertThat(orderTable.isEmpty()).isFalse();
     }
 
     @DisplayName("손님의 수가 -1인 OrderTable을 생성한다.")
@@ -36,7 +36,7 @@ class OrderTableTest {
 
         orderTable.changeEmpty(true);
 
-        assertThat(orderTable.isEmpty()).isTrue();
+        Assertions.assertThat(orderTable.isEmpty()).isTrue();
     }
 
     @DisplayName("테이블의 손님 수를 변경한다.")
@@ -46,7 +46,7 @@ class OrderTableTest {
 
         orderTable.changeNumberOfGuests(5);
 
-        assertThat(orderTable.getNumberOfGuests().value()).isEqualTo(5);
+        Assertions.assertThat(orderTable.getNumberOfGuests().value()).isEqualTo(5);
     }
 
     @DisplayName("테이블의 손님 수를 변경한다.")
