+++ kitchenpos-table/src/test/java/kitchenpos/table/domain/TableGroupTest.java
@@ -1,12 +1,11 @@
 package kitchenpos.table.domain;
 
+import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.List;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 class TableGroupTest {
     private final List<OrderTable> orderTables = Arrays.asList(new OrderTable(0, true),
@@ -17,7 +16,7 @@ class TableGroupTest {
     @Test
     void create() {
         TableGroup tableGroup = new TableGroup();
-        assertThat(tableGroup.getCreatedDate()).isNotNull();
+        Assertions.assertThat(tableGroup.getCreatedDate()).isNotNull();
     }
 
     @DisplayName("단체 지정에 테이블들을 추가한다.")
@@ -27,7 +26,7 @@ class TableGroupTest {
 
         tableGroup.addOrderTables(orderTables);
 
-        assertThat(tableGroup.getOrderTables()).hasSize(3);
+        Assertions.assertThat(tableGroup.getOrderTables()).hasSize(3);
     }
 
     @DisplayName("단체 지정을 해지한다.")
@@ -38,6 +37,6 @@ class TableGroupTest {
 
         tableGroup.ungroup();
 
-        assertThat(tableGroup.getOrderTables()).isEmpty();
+        Assertions.assertThat(tableGroup.getOrderTables()).isEmpty();
     }
 }
