+++ settings.gradle
@@ -2,3 +2,5 @@ rootProject.name = 'kitchenpos'
 include 'kitchenpos-menu'
 include 'kitchenpos-order'
 include 'kitchenpos-product'
+include 'kitchenpos-table'
+include 'kitchenpos-common'
